requirements.txt
motor
aiohttp
requests
python-telegram-bot==20.6
pymongo 
pyrate-limiter==3.1.0
apscheduler==3.6.3
pyrogram
python-dotenv
cachetools
dclua
